cli/api: Split handlers by resource
