+++ packages/cli/src/api/__tests__/bookmark-handlers.test.ts
@@ -1,249 +1,37 @@
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-
-const coreMocks = vi.hoisted(() => ({
-  deleteBookmark: vi.fn(),
-  deleteSessionAlias: vi.fn(),
-  importBookmarks: vi.fn(),
-  listBookmarks: vi.fn(),
-  listDashboardCostFacts: vi.fn(() => null),
-  listFileActivity: vi.fn(),
-  listSessionAliases: vi.fn(),
-  upsertBookmark: vi.fn(),
-  upsertSessionAlias: vi.fn(),
-}));
-
-const loggerMocks = vi.hoisted(() => ({
-  debug: vi.fn(),
-  error: vi.fn(),
-  info: vi.fn(),
-  warn: vi.fn(),
-}));
-
-vi.mock("@codesesh/core/runtime/state", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("@codesesh/core/runtime/state")>();
-  return {
-    ...actual,
-    deleteBookmark: coreMocks.deleteBookmark,
-    deleteSessionAlias: coreMocks.deleteSessionAlias,
-    importBookmarks: coreMocks.importBookmarks,
-    listBookmarks: coreMocks.listBookmarks,
-    listSessionAliases: coreMocks.listSessionAliases,
-    upsertBookmark: coreMocks.upsertBookmark,
-    upsertSessionAlias: coreMocks.upsertSessionAlias,
-  };
-});
-
-vi.mock("@codesesh/core/runtime/discovery", async (importOriginal) => ({
-  ...(await importOriginal<typeof import("@codesesh/core/runtime/discovery")>()),
-  listDashboardCostFacts: coreMocks.listDashboardCostFacts,
-  listFileActivity: coreMocks.listFileActivity,
-}));
-
-vi.mock("../../logging.js", () => ({ appLogger: loggerMocks }));
-
+import { describe, expect, it, vi } from "vitest";
+import type {
+  BookmarkView,
+  LiveSnapshot,
+  ScanResultSource,
+} from "./state-handler-test-fixtures.js";
 import {
+  availableBookmark,
+  bookmarkScanSource,
+  coreMocks,
+  getResponsePayload,
+  legacyBookmark,
+  loggerMocks,
+  makeContext,
+  scanSource,
+  sessionHead,
+  storedBookmark,
+  validReference,
+} from "./state-handler-test-fixtures.js";
+
+const {
   BookmarkStorageUnavailableError,
   SessionAliasValidationError,
   StateStorageUnavailableError,
-  type BookmarkRecord,
-  type BookmarkView,
-} from "@codesesh/core/runtime/state";
-import { type LiveSnapshot, type SessionHead } from "@codesesh/core/runtime/discovery";
-import {
-  handleGetAgents,
-  handleGetDashboard,
-  handleGetFileActivity,
-  handleGetProjects,
-  handleGetSessions,
-  handlePostClientLog,
-  handleSearchSessions,
-  type ScanResultSource,
-} from "../handlers.js";
-import {
+} = await import("@codesesh/core/runtime/state");
+const {
   handleDeleteBookmark,
   handleDeleteSessionAlias,
   handleGetBookmarks,
   handleImportBookmarks,
   handlePutBookmark,
   handlePutSessionAlias,
-} from "../bookmark-handlers.js";
-import { addCalendarDays } from "@codesesh/core/contract";
-import { createApiRoutes } from "../routes.js";
-import { invalidateAliasView } from "../session-aliases-view.js";
-import type { ProjectIdentityResolver } from "../../project-identity-resolver.js";
-
-interface ContextOptions {
-  body?: unknown;
-  rejectBody?: boolean;
-  param?: Record<string, string>;
-  query?: Record<string, string>;
-}
-
-function makeContext(options: ContextOptions = {}) {
-  const params = new URLSearchParams(options.query ?? {});
-  const url = `http://localhost/${params.size > 0 ? `?${params.toString()}` : ""}`;
-  return {
-    req: {
-      json: () =>
-        options.rejectBody
-          ? Promise.reject(new SyntaxError("invalid JSON"))
-          : Promise.resolve(options.body),
-      param: (key: string) => options.param?.[key],
-      query: (key: string) => options.query?.[key],
-      url,
-      raw: new Request(url),
-    },
-    json: vi.fn((payload: unknown, status = 200) => ({ payload, status })),
-  };
-}
-
-function makeProjectIdentityResolver(): ProjectIdentityResolver {
-  return {
-    resolve: vi.fn(async (cwd: string) => ({
-      identity: { kind: "path" as const, key: cwd, displayName: "repo" },
-      resolverRevision: "project-identity-v2",
-      inputSignature: "test",
-    })),
-    shutdown: vi.fn(async () => {}),
-  };
-}
-
-function getResponsePayload<T>(context: ReturnType<typeof makeContext>): T {
-  return context.json.mock.calls[0]![0] as T;
-}
-
-const validReference = { agentName: "codex", sessionId: "s1" };
-
-const sessionHead: SessionHead = {
-  reference: validReference,
-  title: "Session one",
-  directory: "/workspace",
-  time_created: 1,
-  time_updated: 2,
-  stats: {
-    message_count: 1,
-    total_input_tokens: 2,
-    total_output_tokens: 3,
-    total_cost: 0.1,
-    total_tokens: 5,
-  },
-};
-
-const legacyBookmark = {
-  agentKey: "codex",
-  sessionId: "s1",
-  fullPath: "stale/legacy-route",
-  title: "Session one",
-  directory: "/workspace",
-  time_created: 1,
-  time_updated: 2,
-  bookmarked_at: 3,
-  stats: {
-    message_count: 1,
-    total_input_tokens: 2,
-    total_output_tokens: 3,
-    total_cost: 0.1,
-    total_tokens: 5,
-  },
-};
-
-const storedBookmark: BookmarkRecord = {
-  reference: validReference,
-  bookmarkedAt: 3,
-};
-
-const availableBookmark: BookmarkView = {
-  ...storedBookmark,
-  availability: "available",
-  session: sessionHead,
-};
-
-const scanSource: ScanResultSource = {
-  getSnapshot: () =>
-    ({
-      sessions: [],
-      byAgent: { codex: [] },
-      agents: [],
-    }) as LiveSnapshot,
-};
-
-const bookmarkScanSource: ScanResultSource = {
-  getSnapshot: () =>
-    ({
-      sessions: [sessionHead],
-      byAgent: { codex: [sessionHead] },
-      agents: [],
-    }) as LiveSnapshot,
-};
-
-beforeEach(() => {
-  vi.resetAllMocks();
-  // Successful alias reads are cached for the process lifetime; each test needs
-  // a clean slate or it inherits the previous test's listSessionAliases stub.
-  invalidateAliasView();
-  coreMocks.listBookmarks.mockReturnValue([]);
-  coreMocks.listFileActivity.mockReturnValue([]);
-  coreMocks.listSessionAliases.mockReturnValue([]);
-  coreMocks.upsertBookmark.mockReturnValue(storedBookmark);
-  coreMocks.importBookmarks.mockReturnValue([storedBookmark]);
-  coreMocks.upsertSessionAlias.mockReturnValue({
-    reference: { agentName: "codex", sessionId: "s1" },
-    alias: "Renamed",
-    updatedAt: 1,
-  });
-});
-
-afterEach(() => vi.useRealTimers());
-
-describe("client logging handler", () => {
-  it("rejects malformed and blank log events", async () => {
-    const malformed = makeContext({ rejectBody: true });
-    await handlePostClientLog(malformed as never);
-    expect(malformed.json).toHaveBeenCalledWith({ ok: false }, 400);
-
-    const blank = makeContext({ body: { event: "   " } });
-    await handlePostClientLog(blank as never);
-    expect(blank.json).toHaveBeenCalledWith({ ok: false }, 400);
-    expect(loggerMocks.info).not.toHaveBeenCalled();
-  });
-
-  it("sanitizes event names and bounds structured log data", async () => {
-    const data = {
-      text: "x".repeat(400),
-      count: 2,
-      enabled: true,
-      empty: null,
-      nested: { value: 1 },
-      ...Object.fromEntries(Array.from({ length: 30 }, (_, index) => [`extra${index}`, index])),
-    };
-    const c = makeContext({
-      body: { event: ` feature launch!?${"z".repeat(140)}`, data },
-    });
-
-    await handlePostClientLog(c as never);
-
-    const [event, loggedData] = loggerMocks.info.mock.calls[0]!;
-    expect(event).toMatch(/^client\.feature_launch__/);
-    expect(event).toHaveLength(127);
-    expect(loggedData).toMatchObject({
-      text: "x".repeat(300),
-      count: 2,
-      enabled: true,
-      empty: null,
-      nested: "[object Object]",
-    });
-    expect(Object.keys(loggedData)).toHaveLength(30);
-    expect(c.json).toHaveBeenCalledWith({ ok: true });
-  });
-
-  it("drops non-record log data", async () => {
-    const c = makeContext({ body: { event: "ready", data: "not-an-object" } });
-
-    await handlePostClientLog(c as never);
-
-    expect(loggerMocks.info).toHaveBeenCalledWith("client.ready", {});
-  });
-});
+} = await import("../bookmark-handlers.js");
+const { createApiRoutes } = await import("../routes.js");
 
 describe("bookmark handlers", () => {
   it("materializes current session data and decorates it with aliases", () => {
@@ -663,258 +451,5 @@ describe("session alias handlers", () => {
     expect(() =>
       handleDeleteSessionAlias(makeContext({ param: { agent: "codex", id: "s1" } }) as never),
     ).toThrow("unexpected");
-  });
-});
-
-describe("query boundary handlers", () => {
-  const dateHandlers: Array<{
-    endpoint: string;
-    invoke: (context: ReturnType<typeof makeContext>) => unknown;
-  }> = [
-    { endpoint: "agents", invoke: (context) => handleGetAgents(context as never, scanSource) },
-    { endpoint: "projects", invoke: (context) => handleGetProjects(context as never, scanSource) },
-    { endpoint: "sessions", invoke: (context) => handleGetSessions(context as never, scanSource) },
-    { endpoint: "search", invoke: (context) => handleSearchSessions(context as never, scanSource) },
-    { endpoint: "file-activity", invoke: (context) => handleGetFileActivity(context as never) },
-    {
-      endpoint: "dashboard",
-      invoke: (context) => handleGetDashboard(context as never, scanSource),
-    },
-  ];
-
-  for (const { endpoint, invoke } of dateHandlers) {
-    for (const parameter of ["from", "to"] as const) {
-      it(`rejects an invalid ${parameter} parameter on ${endpoint}`, () => {
-        const context = makeContext({ query: { [parameter]: "not-a-date" } });
-
-        invoke(context);
-
-        expect(context.json).toHaveBeenCalledWith(
-          { error: `${parameter} must be a valid date` },
-          400,
-        );
-        expect(loggerMocks.warn).toHaveBeenCalledWith("api.query_parameter.invalid", {
-          endpoint,
-          parameter,
-          validation_outcome: "rejected",
-        });
-      });
-    }
-  }
-
-  it("reports rejected and empty-result parameter outcomes", () => {
-    const sessions = makeContext({ query: { agent: "nonexistent" } });
-    handleGetSessions(sessions as never, scanSource);
-    const files = makeContext({ query: { limit: "1.5" } });
-    handleGetFileActivity(files as never);
-
-    expect(loggerMocks.warn).toHaveBeenCalledWith("api.query_parameter.invalid", {
-      endpoint: "sessions",
-      parameter: "agent",
-      validation_outcome: "empty_result",
-    });
-    expect(loggerMocks.warn).toHaveBeenCalledWith("api.query_parameter.invalid", {
-      endpoint: "file-activity",
-      parameter: "limit",
-      validation_outcome: "rejected",
-    });
-    expect(files.json).toHaveBeenCalledWith({ error: "limit must be a positive integer" }, 400);
-    expect(coreMocks.listFileActivity).not.toHaveBeenCalled();
-  });
-
-  it("rejects invalid project identities consistently", () => {
-    const sessions = makeContext({ query: { projectKind: "path" } });
-    handleGetSessions(sessions as never, scanSource);
-    expect(sessions.json).toHaveBeenCalledWith(
-      { error: "projectKind and projectKey must form a valid project identity" },
-      400,
-    );
-
-    const files = makeContext({ query: { projectKind: "invalid", projectKey: "/repo" } });
-    handleGetFileActivity(files as never);
-    expect(files.json).toHaveBeenCalledWith(
-      { error: "projectKind and projectKey must form a valid project identity" },
-      400,
-    );
-
-    const dashboard = makeContext({ query: { projectKey: "/repo" } });
-    handleGetDashboard(dashboard as never, scanSource);
-    expect(dashboard.json).toHaveBeenCalledWith(
-      { error: "projectKind and projectKey must form a valid project identity" },
-      400,
-    );
-  });
-
-  it("normalizes file activity filters and caps the result limit", async () => {
-    const c = makeContext({
-      query: {
-        agent: " CoDeX ",
-        sessionId: " s1 ",
-        projectKind: "path",
-        projectKey: "/repo",
-        project: " repo ",
-        cwd: " /repo ",
-        path: " src/index.ts ",
-        kind: "edit",
-        from: "2026-01-01T00:00:00.000Z",
-        to: "2026-01-02T00:00:00.000Z",
-        limit: "999",
-      },
-    });
-
-    const resolver = makeProjectIdentityResolver();
-    await handleGetFileActivity(c as never, {}, resolver);
-
-    expect(coreMocks.listFileActivity).toHaveBeenCalledWith({
-      agent: "codex",
-      sessionId: "s1",
-      projectKind: "path",
-      projectKey: "/repo",
-      project: "repo",
-      projectScope: {
-        identity: { kind: "path", key: "/repo" },
-        path: "/repo",
-      },
-      path: "src/index.ts",
-      kind: "edit",
-      from: new Date("2026-01-01T00:00:00.000Z").getTime(),
-      to: new Date("2026-01-02T00:00:00.000Z").getTime(),
-      limit: 200,
-    });
-    expect(resolver.resolve).toHaveBeenCalledWith("/repo", expect.any(AbortSignal));
-  });
-
-  it.each(["1.5", "0", "-1", "", "Infinity", "invalid"])(
-    "rejects invalid file activity limit %j before storage",
-    (limit) => {
-      const c = makeContext({ query: { kind: "execute", limit } });
-
-      handleGetFileActivity(c as never);
-
-      expect(c.json).toHaveBeenCalledWith({ error: "limit must be a positive integer" }, 400);
-      expect(coreMocks.listFileActivity).not.toHaveBeenCalled();
-    },
-  );
-
-  it("returns an empty file activity result for an unknown agent without querying storage", () => {
-    const c = makeContext({ query: { agent: "nonexistent" } });
-
-    handleGetFileActivity(c as never);
-
-    expect(c.json).toHaveBeenCalledWith({ activity: [] });
-    expect(coreMocks.listFileActivity).not.toHaveBeenCalled();
-  });
-
-  // `days` is the number of calendar days covered, so it matches the bucket count.
-  it("derives dashboard days from custom and default windows", () => {
-    const custom = makeContext({
-      query: {
-        from: "2026-01-01T00:00:00.000Z",
-        to: "2026-01-03T00:00:00.000Z",
-        days: "1",
-      },
-    });
-    handleGetDashboard(custom as never, scanSource);
-    const customFrom = new Date("2026-01-01T00:00:00.000Z").getTime();
-    expect(getResponsePayload<{ window: unknown }>(custom).window).toEqual({
-      from: customFrom,
-      to: new Date("2026-01-03T00:00:00.000Z").getTime(),
-      days: 3,
-      compareFrom: addCalendarDays(customFrom, -3),
-      compareTo: customFrom - 1,
-    });
-
-    const fallback = makeContext({ query: { to: "2026-01-04T00:00:00.000Z" } });
-    handleGetDashboard(fallback as never, scanSource, {
-      from: new Date("2026-01-01T00:00:00.000Z").getTime(),
-    });
-    expect(getResponsePayload<{ window: { days: number } }>(fallback).window.days).toBe(4);
-  });
-
-  it("supports explicit all-time dashboard queries", () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date("2026-01-03T00:00:00.000Z"));
-    const c = makeContext({ query: { days: "0" } });
-
-    handleGetDashboard(c as never, scanSource);
-
-    expect(getResponsePayload<{ window: unknown }>(c).window).toEqual({
-      from: undefined,
-      to: Date.now(),
-      days: 0,
-    });
-  });
-});
-
-describe("session alias caching", () => {
-  const aliasRecord = {
-    reference: { agentName: "codex", sessionId: "s1" },
-    alias: "Renamed",
-    updatedAt: 1,
-  };
-
-  it("queries alias storage once across repeated reads", () => {
-    coreMocks.listSessionAliases.mockReturnValue([aliasRecord]);
-
-    handleGetSessions(makeContext() as never, scanSource);
-    handleGetSessions(makeContext() as never, scanSource);
-    handleGetBookmarks(makeContext() as never, bookmarkScanSource);
-
-    expect(coreMocks.listSessionAliases).toHaveBeenCalledTimes(1);
-  });
-
-  it("picks up a stored alias on the next read", async () => {
-    coreMocks.listSessionAliases.mockReturnValue([]);
-    handleGetSessions(makeContext() as never, scanSource);
-
-    coreMocks.upsertSessionAlias.mockReturnValue(aliasRecord);
-    coreMocks.listSessionAliases.mockReturnValue([aliasRecord]);
-    await handlePutSessionAlias(
-      makeContext({ body: { alias: "Renamed" }, param: { agent: "codex", id: "s1" } }) as never,
-    );
-
-    coreMocks.listBookmarks.mockReturnValue([storedBookmark]);
-    const after = makeContext();
-    handleGetBookmarks(after as never, bookmarkScanSource);
-
-    expect(getResponsePayload<{ bookmarks: BookmarkView[] }>(after).bookmarks[0]).toMatchObject({
-      session: { display_title: "Renamed" },
-    });
-  });
-
-  it("drops a removed alias on the next read", () => {
-    coreMocks.listSessionAliases.mockReturnValue([aliasRecord]);
-    coreMocks.listBookmarks.mockReturnValue([storedBookmark]);
-    handleGetBookmarks(makeContext() as never, bookmarkScanSource);
-
-    coreMocks.listSessionAliases.mockReturnValue([]);
-    handleDeleteSessionAlias(makeContext({ param: { agent: "codex", id: "s1" } }) as never);
-
-    const after = makeContext();
-    handleGetBookmarks(after as never, bookmarkScanSource);
-
-    const bookmark = getResponsePayload<{ bookmarks: BookmarkView[] }>(after).bookmarks[0];
-    expect(bookmark).toMatchObject({ availability: "available" });
-    expect(
-      bookmark?.availability === "available" ? bookmark.session : undefined,
-    ).not.toHaveProperty("display_title");
-  });
-
-  it("retries an unavailable store and caches the recovered view", () => {
-    coreMocks.listSessionAliases.mockImplementationOnce(() => {
-      throw new StateStorageUnavailableError();
-    });
-    coreMocks.listSessionAliases.mockReturnValue([aliasRecord]);
-
-    handleGetSessions(makeContext() as never, scanSource);
-    coreMocks.listBookmarks.mockReturnValue([storedBookmark]);
-    const recovered = makeContext();
-    handleGetBookmarks(recovered as never, bookmarkScanSource);
-    handleGetSessions(makeContext() as never, scanSource);
-
-    expect(getResponsePayload<{ bookmarks: BookmarkView[] }>(recovered).bookmarks[0]).toMatchObject(
-      { session: { display_title: "Renamed" } },
-    );
-    expect(coreMocks.listSessionAliases).toHaveBeenCalledTimes(2);
   });
 });
