+++ packages/cli/src/api/client-log-handler.ts
@@ -1,0 +1,24 @@
+import type { Context } from "hono";
+import { appLogger } from "../logging.js";
+import { sanitizeClientLogData } from "./request-payloads.js";
+
+interface ClientLogPayload {
+  event?: unknown;
+  data?: unknown;
+}
+
+export async function handlePostClientLog(c: Context) {
+  const payload = (await c.req.json().catch(() => null)) as ClientLogPayload | null;
+  const rawEvent = payload?.event;
+
+  if (typeof rawEvent !== "string" || !rawEvent.trim()) {
+    return c.json({ ok: false }, 400);
+  }
+
+  const event = rawEvent
+    .trim()
+    .replace(/[^a-zA-Z0-9_.:-]/g, "_")
+    .slice(0, 120);
+  appLogger.info(`client.${event}`, sanitizeClientLogData(payload?.data));
+  return c.json({ ok: true });
+}
